1.7.0版本 增添简单视奸功能

+++ src/main/java/com/duskrainfall/betterminecart/bean/SpringBlock_Data.java
@@ -4,11 +4,11 @@ import java.io.Serializable;
 
 public class SpringBlock_Data implements Serializable {
     public String world;
-    public double x;
-    public double y;
-    public double z;
+    public int x;
+    public int y;
+    public int z;
 
-    public SpringBlock_Data(String world, double x, double y, double z) {
+    public SpringBlock_Data(String world, int x, int y, int z) {
         this.world = world;
         this.x = x;
         this.y = y;
